+++ query/tests/it/interpreters/mod.rs
@@ -28,6 +28,7 @@ mod interpreter_stage_create;
 mod interpreter_stage_drop;
 mod interpreter_table_create;
 mod interpreter_table_drop;
+mod interpreter_table_optimize;
 mod interpreter_table_truncate;
 mod interpreter_use_database;
 mod interpreter_user_alter;
